Refactor EmailInvitationConcern into smaller files
- Splits up EmailInvitationConcern into:
    ParseInvitation
    ProcessInvitation
    EmailInvitation

This will make it easier to read and modify code in the associated service.
