+++ app/services/concerns/course/user_invitation_service/parse_invitation_concern.rb
@@ -1,0 +1,98 @@
+# frozen_string_literal: true
+require 'csv'
+
+# This concern includes methods required to parse the invitations data.
+# This can either be from a form, or a CSV file.
+class Course::UserInvitationService; end
+module Course::UserInvitationService::ParseInvitationConcern
+  extend ActiveSupport::Autoload
+
+  private
+
+  # Invites users to the given course.
+  #
+  # @param [Array<Hash>|File|TempFile] users Invites the given users.
+  # @return [Array<Hash{Symbol=>String}>]
+  #   A mutable array of users to add. Each hash must have three attributes:
+  #     the +:name+,
+  #     the +:email+ of the user to add, as well as
+  #     the intended +:role+ in the course.
+  #   The provided +emails+ are NOT case sensitive.
+  # @raise [CSV::MalformedCSVError] When the file provided is invalid.
+  def parse_invitations(users)
+    result =
+      if users.is_a?(File) || users.is_a?(Tempfile)
+        parse_from_file(users)
+      else
+        parse_from_form(users)
+      end
+    result.each { |user| user[:email] = user[:email].downcase }
+  end
+
+  # Invites the users from the form submission, which reflects the actual model associations.
+  #
+  # We do not use this format in the service object because it is very clumsy.
+  #
+  # @param [Hash] users The attributes from the client.
+  # @return [Array<Hash>] Array of users to be invited
+  def parse_from_form(users)
+    users.map do |(_, value)|
+      { name: value[:name], email: value[:email], role: value[:role] }
+    end
+  end
+
+  # Loads the given file, and entries with blanks in either fields are ignored.
+  # The first row is ignored if it's a header row (contains "name, email"),
+  # else it's treated like a row of student data.
+  #
+  # @param [File] file Reads the given file, in CSV format, for the name and email.
+  # @return [Array<Hash>] The array of records read from the file.
+  # @raise [CSV::MalformedCSVError] When the file provided is invalid.
+  def parse_from_file(file)
+    [].tap do |invites|
+      CSV.foreach(file).with_index(1) do |row, row_number|
+        # Ignore first row if it's a header row.
+        next if row_number == 1 && header_row?(row)
+
+        invite = parse_file_row(row)
+        invites << invite if invite
+      end
+    end
+  rescue StandardError => error
+    raise CSV::MalformedCSVError.new(error), error.message
+  end
+
+  # Returns a boolean to determine whether the row is a header row.
+  #
+  # @param[Array] row Array read from CSV file.
+  # @return [Boolean] Whether the row is a header row
+  def header_row?(row)
+    row[0].casecmp('Name') == 0 && row[1].casecmp('Email') == 0
+  end
+
+  # Parses the given CSV row (array) and returns attributes for a user invitation.
+  #   - Sets the name as the given email if a name was not provided.
+  #
+  # @param [Array] row Array with 3 parameters: name, email and role respectively.
+  # @return [Hash] The parsed invitation attributes given the row.
+  def parse_file_row(row)
+    return nil if row[1].blank?
+    row[0] = row[1] if row[0].blank?
+
+    role = parse_file_role(row[2])
+    { name: row[0], email: row[1], role: role }
+  end
+
+  # Parses the role column from the CSV file.
+  # This method handles the case where the role is not specified too, where "student" will be assumed.
+  #
+  # @param [String] role The role as specified in the CSV file
+  # @return [Integer] The enum integer for +Course::UserInvitation.role+ matching the input.
+  #                   (+Course::UserInvitation.roles[:student]+) is returned by default.
+  def parse_file_role(role)
+    return Course::UserInvitation.roles[:student] if role.blank?
+
+    symbol = role.parameterize(separator: '_').to_sym
+    Course::UserInvitation.roles[symbol] || Course::UserInvitation.roles[:student]
+  end
+end
